refactor: fix typos and minor refactoring
