+++ src/ditto/contexts.py
@@ -28,6 +28,7 @@ from peft import PeftModel
 from torch.fx.experimental.sym_node import SymNode
 from torch.fx.experimental.symbolic_shapes import log as symbolic_shape_logger
 
+from .constants import LOG_LEVEL_LITERAL_MAP
 from .literals import LogLevelLiteral
 from .types import verify
 
@@ -172,11 +173,11 @@ def set_logger_level(logger_name: str, log_level: LogLevelLiteral) -> Generator[
 
     Args:
         logger_name (str): The name of the logger to modify
-        log_level (int): The log level to temporarily set.
+        log_level (LogLevelLiteral): The log level to temporarily set.
     """
     internal_logger = logging.getLogger(logger_name)
     original_level = internal_logger.level
-    internal_logger.setLevel(log_level)
+    internal_logger.setLevel(LOG_LEVEL_LITERAL_MAP[log_level])
     try:
         yield None
     finally:
