+++ src/ditto/fx/passes/__init__.py
@@ -48,7 +48,7 @@ from .parallelize_pipeline import ParallelizePipeline
 from .pop_lora_plugins import PopLoraPlugins
 from .propagate_tensor_parallelism import PropagateTensorParallelism
 from .replace_mm_by_gemm_plugin import ReplaceMMByGemmPlugin
-from .replace_moe_by_mixture_or_experts_plugin import ReplaceMoEByMoEPlugin
+from .replace_moe_by_mixture_of_experts_plugin import ReplaceMoEByMoEPlugin
 from .replace_sdpa_by_gpt_attention_plugin import ReplaceSDPAByGPTAttentionPlugin
 from .replace_view_by_reshape import ReplaceViewByReshape
 from .reset_code_gen import ResetCodeGen
